% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_model_simulator}
\alias{create_model_simulator}
\title{Create a simulator function of a GAM}
\usage{
create_model_simulator(model, post = identity)
}
\arguments{
\item{model}{}

\item{post}{a \code{function(x) {...}} applied to the output of the resulting
function}
}
\value{
a closure of \code{newdata} and \code{newparms} that returns a vector of simulated
data from the fitted GAM model
}
\description{
Create a simulator function of a GAM
}
