% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_sim_basis_maker}
\alias{create_sim_basis_maker}
\title{Creates a function which can produce the basis for simulation datasets}
\usage{
create_sim_basis_maker(basisdt)
}
\arguments{
\item{basisdt}{a dataset containing \code{pid}, \code{geometry}, \code{neighbor_pids},
\code{neighbor_boundary_lengths}}
}
\description{
Creates a function which can produce the basis for simulation datasets
}
