% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_simulator}
\alias{create_simulator}
\title{Create a simulator}
\usage{
create_simulator(exposure_sim, outcome_sim)
}
\arguments{
\item{exposure_sim}{the result of \code{create_model_simulator} for the exposure
of interest}

\item{outcome_sim}{the result of \code{create_model_simulator} for the outcome
of interest}
}
\description{
Create a simulator
}
