% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{.pw}
\alias{.pw}
\title{Estimate the density ratio by modeling whether an observation is from
the permuted dataset or original dataset}
\usage{
.pw(dt, rhs_formula, fitter)
}
\description{
Estimate the density ratio by modeling whether an observation is from
the permuted dataset or original dataset
}
