% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{.shared_boundaries_length}
\alias{.shared_boundaries_length}
\title{Computes the length of shared boundaries between a parcel and its neighbors}
\usage{
.shared_boundaries_length(x, neighbors)
}
\arguments{
\item{x}{the \code{MULTIPOLYGON} for the parcel of interest}

\item{neighbors}{a list of \code{MULTIPOLYGON}s of which share a boundary with \code{x}}
}
\description{
Computes the length of shared boundaries between a parcel and its neighbors
}
\keyword{internal}
