% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{get_permutation_weights}
\alias{get_permutation_weights}
\title{Estimate permutation weights B times and average the results}
\usage{
get_permutation_weights(
  dt,
  B,
  rhs_formula,
  fitter,
  modify_C = identity,
  predictor,
  fitter_args = list()
)
}
\description{
Estimate permutation weights B times and average the results
}
