% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{make_ipw_estimator}
\alias{make_ipw_estimator}
\title{Make an ipw estimator}
\usage{
make_ipw_estimator(
  fitter,
  rhs_formula,
  predictor = function(object, newdata) {     predict(object = object, newdata =
    newdata, type = "response") },
  fitter_args = list()
)
}
\description{
Make an ipw estimator
}
