% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{make_pw_estimator}
\alias{make_pw_estimator}
\title{Create a permutation weighted estimator for the marginal structural model}
\usage{
make_pw_estimator(fitter, rhs_formula, B, msm_formula = Y ~ A * A_tilde)
}
\description{
Create a permutation weighted estimator for the marginal structural model
}
