% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{make_sim}
\alias{make_sim}
\title{Create a simulation dataset}
\usage{
make_sim(
  basedt,
  simulator,
  parms = list(exposure_newparms = identity, outcome_newparms = identity)
)
}
\description{
Create a simulation dataset
}
