% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{make_sim_basis_data}
\alias{make_sim_basis_data}
\title{Create a simulation basis dataset}
\usage{
make_sim_basis_data(
  basis_maker,
  id,
  buffer,
  checker = function(hash) {     FALSE },
  redirector = function(out, hash) {     out }
)
}
\description{
Create a simulation basis dataset
}
