% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{permute_A}
\alias{permute_A}
\title{Permutes the treatment vector and recomputes the proportion of neighbors
treated.}
\usage{
permute_A(dt)
}
\description{
Permutes the treatment vector and recomputes the proportion of neighbors
treated.
}
