% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_permuation.R
\name{permute_and_stack}
\alias{permute_and_stack}
\title{Stack the observed data with the permuted data}
\usage{
permute_and_stack(dt)
}
\description{
Stack the observed data with the permuted data
}
