% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{shared_boundaries_length}
\alias{shared_boundaries_length}
\title{Computes the length of shared boundaries between a parcel and its neighbors}
\usage{
shared_boundaries_length(df, i, ngbrs, boundary_col = "boundary")
}
\arguments{
\item{df}{a spatial dataframe, containing a variable named by \code{boundary_col}
(defaulting to \code{"boundary"}) which is a \code{LINES} object}

\item{i}{the row index in \code{df} of the parcel of interest}

\item{nghrs}{the indices of neighbors of \code{df[i, ]} in \code{df}}
}
\value{
a \code{units} vector containing boundary lengths in meters
}
\description{
Computes the length of shared boundaries between a parcel and its neighbors
}
